rename and remove:  - storage dir to data dir  - Mappers into NewsItemMapper  - replace NewsItemMapper into storage/mappers/
