+++ app/src/main/java/ru/vssemikoz/newsfeed/data/NewsApiRepository.java
@@ -1,4 +1,4 @@
-package ru.vssemikoz.newsfeed.storage;
+package ru.vssemikoz.newsfeed.data;
 
 import org.jetbrains.annotations.NotNull;
