+++ app/src/main/java/ru/vssemikoz/newsfeed/data/mappers/NewsItemMapper.java
@@ -1,4 +1,4 @@
-package ru.vssemikoz.newsfeed.utils;
+package ru.vssemikoz.newsfeed.data.mappers;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +10,7 @@ import ru.vssemikoz.newsfeed.models.NewsApiResponse;
 import ru.vssemikoz.newsfeed.models.NewsApiResponseItem;
 import ru.vssemikoz.newsfeed.models.NewsItem;
 
-public class Mappers {
+public class NewsItemMapper {
     public static List<NewsItem> mapResponseToNewsItems(Response<NewsApiResponse> response, Category category) {
         List<NewsItem> news = new ArrayList<>();
         List<NewsApiResponseItem> responseItems = Objects.requireNonNull(response.body()).getNewsApiResponseItemList();
