+++ app/src/main/java/ru/vssemikoz/newsfeed/newsfeed/NewsFeedFragment.java
@@ -23,7 +23,7 @@ import ru.vssemikoz.newsfeed.adapters.NewsFeedAdapter;
 import ru.vssemikoz.newsfeed.dialogs.PickCategoryDialog;
 import ru.vssemikoz.newsfeed.models.Category;
 import ru.vssemikoz.newsfeed.models.NewsItem;
-import ru.vssemikoz.newsfeed.storage.IconicStorage;
+import ru.vssemikoz.newsfeed.data.IconicStorage;
 
 import static androidx.core.util.Preconditions.checkNotNull;
